Make RangeSlider.Preview muted version of the original
Because RangeSlider's series colour settings are directly linked to the graphs
they represent, there's no way to use a different colour pallete, say,
grayscale, to mute the preview to define it as separate to the original.

This change sets the opacity to give a muted colour view to the range slider
preview.

This could be extended to be a user-definable setting.